refactor(polars): remove casting where possible; handle conversion on output

+++ ibis/backends/polars/tests/test_udf.py
@@ -47,7 +47,7 @@ def test_multiple_argument_udf(alltypes):
     df = alltypes[["smallint_col", "int_col"]].execute()
     expected = df.smallint_col + df.int_col
 
-    tm.assert_series_equal(result, expected.rename("tmp"))
+    tm.assert_series_equal(result, expected.astype("int64").rename("tmp"))
 
 
 @pytest.mark.parametrize(
